Delete team.txt (description below)
We had to remove this file because Roy pulled the repository after the team.txt file was created. So he will now pull the repo without the team.txt file in it to simulate the correct environment. I will add the team.txt back after he pulls the repo with the 3 names in it (excluding his). Then he will get a push error and have to merge properly